new CreateSharedExpense works! Status Codes declared, ExpenseCategoryConverter Added, New Exception Handling implemented in Invite and Expense APIs

+++ WiselySplitServer/src/main/java/ca/mohawkCollege/WiselySplitServer/jpa/entities/ExpenseParticipationId.java
@@ -14,7 +14,7 @@ import java.util.Objects;
 @Data
 public class ExpenseParticipationId implements Serializable {
 
-    private Integer expenseId;
+    private Long expenseId;
     private Long userId;
 
     @Override
